console: cleaned up with consoles now going through main websocket

+++ lib/tab/pty.dart
@@ -16,8 +16,6 @@ class CmdrPty {
   Process _shell;
   String _workspacePath;
   Socket _ptySocket;
-  String _port;
-  StreamController<List<int>> _clientServerInterface;
 
   CmdrPty(this.ptyNum, String workspacePath, String numRows, String numCols) {
     help.debug('Spawning UpDroidPty ($ptyNum)', 0);
@@ -29,9 +27,6 @@ class CmdrPty {
   }
 
   void _startPty(UpDroidMessage um) {
-    print('starting up');
-    _clientServerInterface = new StreamController<List<int>>();
-
     // Process launches 'cmdr-pty', a go program that provides a direct hook to a system pty.
     // See http://bitbucket.org/updroid/cmdr-pty
     Process.start('cmdr-pty', ['-protocol', 'tcp', '-size', '${um.body}'], environment: {'TERM':'vt100'}, workingDirectory: _workspacePath).then((Process shell) {
@@ -44,20 +39,15 @@ class CmdrPty {
       portListener = stdoutBroadcast.listen((data) {
         String dataString = UTF8.decode(data);
         if (dataString.contains('now listening on:  [::]:')) {
-          String _port = dataString.replaceFirst('now listening on:  [::]:', '');
-          print(_port);
-          UpDroidMessage portMessage = new UpDroidMessage('PTY_READY', _port);
+          String port = dataString.replaceFirst('now listening on:  [::]:', '');
+          UpDroidMessage portMessage = new UpDroidMessage('PTY_READY', '');
           mailbox.ws.add(portMessage.s);
           portListener.cancel();
 
-          Socket.connect('127.0.0.1', int.parse(_port)).then((socket) {
+          Socket.connect('127.0.0.1', int.parse(port)).then((socket) {
             _ptySocket = socket;
-            print('Connected to: ${socket.remoteAddress.address}:${socket.remotePort}');
-            socket.listen((data) {
-              print(data);
-//              print(UTF8.decode(data));
-              _clientServerInterface.sink.add((data));
-            });
+            help.debug('CmdrPty-$ptyNum connected to: ${socket.remoteAddress.address}:${socket.remotePort}', 0);
+            socket.listen((data) => mailbox.ws.add((data)));
           });
         }
       });
@@ -73,13 +63,9 @@ class CmdrPty {
   }
 
   void _handleIOStream(HttpRequest request) {
-    print('in handle');
-    mailbox.ws.where((e) => request.uri.path == '/${guiName.toLowerCase()}/$ptyNum/cmdr-pty').listen((data) {
-//      print(UTF8.decode(data));
-      _ptySocket.add(data);
-    });
-
-    _clientServerInterface.stream.listen((data) => mailbox.ws.add((data)));
+    mailbox.ws
+      .where((e) => request.uri.path == '/${guiName.toLowerCase()}/$ptyNum/cmdr-pty')
+      .listen((data) => _ptySocket.add(data));
   }
 
   void _resize(UpDroidMessage um) {
@@ -87,9 +73,8 @@ class CmdrPty {
   }
 
   void cleanup() {
-    _clientServerInterface.close();
-    _ptySocket.destroy();
-    _shell.kill();
+    if (_ptySocket != null) _ptySocket.destroy();
+    if (_shell != null) _shell.kill();
   }
 
   void _registerMailbox() {
